tech-debt: split blizzard.service.ts — extract BlizzardAuthService + helpers (ROK-1180)
