+++ api/src/plugins/wow-common/blizzard.service.ts
@@ -1,7 +1,4 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { OnEvent } from '@nestjs/event-emitter';
-import { SettingsService } from '../../settings/settings.service';
-import { SETTINGS_EVENTS } from '../../settings/settings.types';
 import { memorySwr } from '../../common/swr-cache';
 import type { WowGameVariant } from '@raid-ledger/contract';
 
@@ -14,19 +11,17 @@ import {
   type WowInstanceDetail,
   type RealmCacheEntry,
   type InstanceListCacheEntry,
-  type InstanceListCacheData,
   type InstanceDetailCacheEntry,
-  SPEC_ROLE_MAP,
-  TOKEN_EXPIRY_BUFFER,
   REALM_CACHE_TTL,
   INSTANCE_CACHE_TTL,
 } from './blizzard.constants';
-import { buildCharacterParams } from './blizzard-character.helpers';
+import { BlizzardAuthService } from './blizzard-auth.service';
 import * as profileH from './blizzard-profile.helpers';
 import * as equipH from './blizzard-equipment.helpers';
 import * as profH from './blizzard-professions.helpers';
 import * as specH from './blizzard-spec.helpers';
 import * as instH from './blizzard-instance.helpers';
+import * as instFetch from './blizzard-instance.fetch';
 import type { ExternalCharacterProfessions } from '../plugin-host/extension-types';
 
 // Re-export types for backward compatibility
@@ -43,30 +38,11 @@ export type {
 @Injectable()
 export class BlizzardService {
   private readonly logger = new Logger(BlizzardService.name);
-  private accessToken: string | null = null;
-  private tokenExpiry: Date | null = null;
-  private tokenFetchPromise: Promise<string> | null = null;
   private realmCache = new Map<string, RealmCacheEntry>();
   private instanceListCache = new Map<string, InstanceListCacheEntry>();
   private instanceDetailCache = new Map<string, InstanceDetailCacheEntry>();
 
-  constructor(private readonly settingsService: SettingsService) {}
-
-  @OnEvent(SETTINGS_EVENTS.BLIZZARD_UPDATED)
-  handleBlizzardConfigUpdate(): void {
-    this.accessToken = null;
-    this.tokenExpiry = null;
-    this.tokenFetchPromise = null;
-    this.logger.log('Blizzard config updated — cached token cleared');
-  }
-
-  normalizeRealmSlug(realm: string): string {
-    return realm.toLowerCase().replace(/'/g, '').replace(/\s+/g, '-').trim();
-  }
-
-  specToRole(spec: string): 'tank' | 'healer' | 'dps' | null {
-    return SPEC_ROLE_MAP[spec] ?? null;
-  }
+  constructor(private readonly auth: BlizzardAuthService) {}
 
   /** Fetch a character profile from the Blizzard API.
    * @param apiNamespacePrefix - From the game row (null for retail)
@@ -77,7 +53,7 @@ export class BlizzardService {
     region: string,
     apiNamespacePrefix: string | null = null,
   ): Promise<BlizzardCharacterProfile> {
-    const token = await this.getAccessToken(region);
+    const token = await this.auth.getAccessToken(region);
     const data = await profileH.fetchProfileData(
       name,
       realm,
@@ -98,16 +74,14 @@ export class BlizzardService {
     );
   }
 
-  /** Fetch character equipment from the Blizzard API.
-   * @param apiNamespacePrefix - From the game row (null for retail)
-   */
+  /** Fetch character equipment from the Blizzard API. */
   async fetchCharacterEquipment(
     name: string,
     realm: string,
     region: string,
     apiNamespacePrefix: string | null = null,
   ): Promise<BlizzardCharacterEquipment | null> {
-    const token = await this.getAccessToken(region);
+    const token = await this.auth.getAccessToken(region);
     return equipH.fetchCharacterEquipment(
       name,
       realm,
@@ -118,16 +92,14 @@ export class BlizzardService {
     );
   }
 
-  /** Fetch character professions from the Blizzard API.
-   * @param apiNamespacePrefix - From the game row (null for retail)
-   */
+  /** Fetch character professions from the Blizzard API. */
   async fetchCharacterProfessions(
     name: string,
     realm: string,
     region: string,
     apiNamespacePrefix: string | null = null,
   ): Promise<ExternalCharacterProfessions | null> {
-    const token = await this.getAccessToken(region);
+    const token = await this.auth.getAccessToken(region);
     return profH.fetchCharacterProfessions(
       name,
       realm,
@@ -138,9 +110,7 @@ export class BlizzardService {
     );
   }
 
-  /** Fetch character specializations from the Blizzard API.
-   * @param apiNamespacePrefix - From the game row (null for retail)
-   */
+  /** Fetch character specializations from the Blizzard API. */
   async fetchCharacterSpecializations(
     name: string,
     realm: string,
@@ -148,30 +118,19 @@ export class BlizzardService {
     characterClass: string,
     apiNamespacePrefix: string | null = null,
   ): Promise<InferredSpecialization> {
-    try {
-      const token = await this.getAccessToken(region);
-      const { realmSlug, charName, namespace, baseUrl } = buildCharacterParams(
-        name,
-        realm,
-        region,
-        apiNamespacePrefix,
-      );
-      const url = `${baseUrl}/profile/wow/character/${realmSlug}/${charName}/specializations?namespace=${namespace}&locale=en_US`;
-      const res = await fetch(url, {
-        headers: { Authorization: `Bearer ${token}` },
-      });
-      if (!res.ok) return specH.NO_SPEC;
-      const data = (await res.json()) as Record<string, unknown>;
-      return specH.parseSpecData(data, characterClass);
-    } catch (err) {
-      this.logger.debug(`Failed to fetch specializations: ${err}`);
-      return specH.NO_SPEC;
-    }
+    const token = await this.auth.getAccessToken(region);
+    return specH.fetchCharacterSpecializations(
+      name,
+      realm,
+      region,
+      characterClass,
+      apiNamespacePrefix,
+      token,
+      this.logger,
+    );
   }
 
-  /** Fetch the realm list for a given region and namespace.
-   * @param apiNamespacePrefix - From the game row (null for retail)
-   */
+  /** Fetch the realm list for a given region and namespace. */
   async fetchRealmList(
     region: string,
     apiNamespacePrefix: string | null = null,
@@ -182,7 +141,7 @@ export class BlizzardService {
       key: cacheKey,
       ttlMs: REALM_CACHE_TTL,
       fetcher: async () => {
-        const token = await this.getAccessToken(region);
+        const token = await this.auth.getAccessToken(region);
         return instH.fetchRealmListFromApi(
           region,
           apiNamespacePrefix,
@@ -201,34 +160,11 @@ export class BlizzardService {
       cache: this.instanceListCache,
       key: `${region}:${gameVariant}`,
       ttlMs: INSTANCE_CACHE_TTL,
-      fetcher: () => this._fetchAllInstancesFromApi(region, gameVariant),
+      fetcher: async () => {
+        const token = await this.auth.getAccessToken(region);
+        return instFetch.fetchAllInstancesFromApi(region, gameVariant, token);
+      },
     });
-  }
-
-  private async _fetchAllInstancesFromApi(
-    region: string,
-    gameVariant: WowGameVariant,
-  ): Promise<InstanceListCacheData> {
-    const token = await this.getAccessToken(region);
-    const tiers = await instH.fetchExpansionIndex(region, token);
-    const details = await instH.fetchExpansionDetails(
-      tiers,
-      `https://${region}.api.blizzard.com`,
-      `static-${region}`,
-      token,
-    );
-    let { dungeons, raids } = instH.mergeExpansionInstances(details);
-    ({ dungeons, raids } = instH.filterByVariant(dungeons, raids, gameVariant));
-    dungeons = instH.deduplicateById(dungeons);
-    raids = instH.deduplicateById(raids);
-    if (gameVariant !== 'retail') {
-      dungeons = instH.expandSubInstances(dungeons);
-      raids = instH.expandSubInstances(raids);
-    }
-    return {
-      dungeons: dungeons.map((i) => instH.enrichInstance(i, gameVariant)),
-      raids: raids.map((i) => instH.enrichInstance(i, gameVariant)),
-    };
   }
 
   async fetchInstanceDetail(
@@ -240,43 +176,23 @@ export class BlizzardService {
       cache: this.instanceDetailCache,
       key: `${region}:${gameVariant}:${instanceId}`,
       ttlMs: INSTANCE_CACHE_TTL,
-      fetcher: () =>
-        this._fetchInstanceDetailFromApi(instanceId, region, gameVariant),
+      fetcher: async () => {
+        const token = await this.auth.getAccessToken(region);
+        return instFetch.fetchInstanceDetailFromApi(
+          instanceId,
+          region,
+          gameVariant,
+          token,
+        );
+      },
     });
-  }
-
-  private async _fetchInstanceDetailFromApi(
-    instanceId: number,
-    region: string,
-    gameVariant: WowGameVariant,
-  ): Promise<WowInstanceDetail> {
-    if (instanceId > 10000) {
-      const synth = instH.resolveSyntheticInstance(instanceId);
-      if (synth) return synth;
-    }
-    const token = await this.getAccessToken(region);
-    const url = `https://${region}.api.blizzard.com/data/wow/journal-instance/${instanceId}?namespace=static-${region}&locale=en_US`;
-    const res = await fetch(url, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    if (!res.ok)
-      throw new Error(`Failed to fetch instance detail (${res.status})`);
-    const data = (await res.json()) as {
-      id: number;
-      name: string;
-      minimum_level?: number;
-      modes?: Array<{ mode: { type: string }; players: number }>;
-      category?: { type: string };
-      expansion?: { name: string };
-    };
-    return instH.buildInstanceDetail(data, gameVariant);
   }
 
   async fetchBlizzardApi<T = unknown>(
     url: string,
     region: string = 'us',
   ): Promise<T | null> {
-    const token = await this.getAccessToken(region);
+    const token = await this.auth.getAccessToken(region);
     const res = await fetch(url, {
       headers: { Authorization: `Bearer ${token}` },
     });
@@ -285,48 +201,5 @@ export class BlizzardService {
       return null;
     }
     return res.json() as Promise<T>;
-  }
-
-  private async getAccessToken(region: string): Promise<string> {
-    if (this.accessToken && this.tokenExpiry && new Date() < this.tokenExpiry)
-      return this.accessToken;
-    if (this.tokenFetchPromise) return this.tokenFetchPromise;
-    this.tokenFetchPromise = this.fetchNewToken(region);
-    try {
-      return await this.tokenFetchPromise;
-    } finally {
-      this.tokenFetchPromise = null;
-    }
-  }
-
-  private async fetchNewToken(region: string): Promise<string> {
-    const config = await this.settingsService.getBlizzardConfig();
-    if (!config) throw new Error('Blizzard API credentials not configured');
-    const response = await fetch(`https://${region}.battle.net/oauth/token`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${Buffer.from(`${config.clientId}:${config.clientSecret}`).toString('base64')}`,
-      },
-      body: new URLSearchParams({ grant_type: 'client_credentials' }),
-    });
-    if (!response.ok) {
-      const errorText = await response.text();
-      this.logger.error(
-        `Failed to get Blizzard access token: ${response.status} ${errorText}`,
-      );
-      throw new Error(
-        `Failed to get Blizzard access token: ${response.statusText}`,
-      );
-    }
-    const data = (await response.json()) as {
-      access_token: string;
-      expires_in: number;
-    };
-    this.accessToken = data.access_token;
-    this.tokenExpiry = new Date(
-      Date.now() + (data.expires_in - TOKEN_EXPIRY_BUFFER) * 1000,
-    );
-    return this.accessToken;
   }
 }
